cns_drone_simulator
The purpose of the "avNET" project is to create and manage routes for aircraft.
In fact, as a result of entering a set of coordinates, a path can be generated and upload it to a MongoDB database. The ultimate goal is to perform a simulation that tests the integration of a vehicle with the data saved on the JSON file in our database.